# Project Setup
Today we're going to be tackling the [Molecule to Atom]({{page.repo}})
kata. The goal is to parse a string representation of a molecule into
its components elements, or atoms.

As always, the first step is to get our project set up. We'll be using
Mocha as our test runner and Babel for ES6 support.