# Laying the Groundwork
The best place to start is at the beginning. We'll begin solving this
kata by writing the most basic test we can think of. We expect an empty
string to be parsed into an empty object:

    expect(parseMolecule("")).to.deep.equal({});

After writing this test, it fails. `parseMolecule` is not defined. We
can quickly remedy this situation by importing `parseMolecule` into our
test file and then exporting it from our main module.

Lastly, we need `parseMolecule` to return an empty object. Just like
that, our tests are green.