# Introducing Our Abstractions
Know this solution is going to get complex fairly quickly, and we'd
like to leverage some abstractions to make it more testable and
maintainable.

The most obvious abstraction we can think of is a `molecule`. We create
a `molecule` by passing in a `formula`. From there, we can call `parse`
to get an object-based representation of the molecule.

We can rewrite our `parseMolecule` function to use `molecule` in a clean
way.