# Testing Hydrogen
Let's add a new test case:

    expect(parseMolecule("H")).to.deep.equal({ H: 1 });

This test fails. It's expecting `{}` to equal `{ H: 1 }`.

The fastest way we can get ourselves to green is to return an object
that maps the provided `formula` to `1`.

    return {
        [formula]: 1
    };

After making this change, our first test fails. We now need to
explicitly handle this base case:

    if (!formula) {
        return {};
    }

And now we're back to green again.