# Compiling Parts
Things are getting more complicated. Instead of just breaking apart our
`formula` and looking for elements, we need to look for pairs of
elements and their corresponding `count`, or ratio in the molecule.

A nice way to do this is using regular expression match groups and ES6
pattern matching. We can define our regex, and wrap the element, and
optional count in match groups:

    let regex = /([A-Z][a-z]*)(\d*)/g;

As we iterate over our regex matches, we can destructure the results
into its corresponding `element` and `count` pair.

Finally, we restructured `parts` to return an object with `part` and
`count` keys to keep things explicit. Thinking ahead, I think this might
come in handy when we need to deal with nested expressions.