#Fixing an Edge Case
One thing I noticed while looking over our solution so far is that it
did not support multiple instances of the same element in a molecule.
For example, with `"HMgH"`, the last `"H"` element would override the
previous.

I added a test to confirm my suspicions, and quickly fixed the problem.
Instead of overriding the element in the `result` object with `count`,
increment it (if it exists) by `count`:

    result[part] = ~~result[part] + count;

The `~~` operator is rarely seen, but it's an incredibly useful
operator. In this case, if `result[part]` is undefined, it will be
converted to `0` instead, preventing a `NaN` result from our addition.