# Multiplying Nested Expressions
With that refactor out of the way, we can easily implement `multiply`.
To give ourselves an explicit objective, let's write a test first:

    expect(molecule("H").multiply(2).parse()).to.deep.equal({ H: 2 });

Implementing `multiply` is straight forward. We just loop over each of
the `_parts` of the molecule, multiplying its `count` by the provided
`count`. Finally, we return `this`, so calls to `molecule` methods can
be chained together.

Next, let's write a test for handling ratios on nested expressions:

    expect(parseMolecule("[HO]2Mg")).to.deep.equal({ H: 2, O: 2, Mg: 1 });

This test fails, as expected, but it's not difficult to get our suite
back to green. We just need to build our `nested` molecule, `multiply`
is by its corresponding `count`, and finally append its `parts` onto our
results array.