# Regex Woes
Unfortunately, there's a big problem with our solution. Our regex
(`\[.*\]`) is looking for opening and closing brackets and assuming that
everything within them are a nested expression.

This assumption breaks down when we have two nested expressions in the
same formula, like `"[H]O[H]"`. Our regex will match on the first and
very last square brackets, returning `"H]O[H"` as the nested expression.
We can write this as a test to see the failure in action:

    expect(molecule("[H]2O[H]").parts()).to.deep.equal([
        { part: "H", count: 2 },
        { part: "O", count: 1 },
        { part: "H", count: 1 }
    ]);

Switching to a non-greedy regex (`\[.*?\]`) would still fail, but this
time on nested subexpressions, like `"[H[O]]"`.

We need a better way to parse out our formula parts.

Our new solution uses a regex to match on all opening brackets, closing
brackets, and elements with a trailing count:

    let regex = /((\[)|(\])|([A-Z][a-z]*))(\d*)/g;

Every time we encounter an opening bracket, we push a new formula onto
our new `stack`:

    if (open) {
        stack.push({
            formula: ""
        });
    }

A non-empty stack means that we're collecting the pieces of a nested
expression. That means we should just append any elements and counts we
find to that sub-expression's formula:

    else if (stack.length) {
        stack[stack.length - 1].formula += part + count;
    }

Finally, whenever we encounter a closing bracket, we want to create a
new `molecule` out of the nested expression we just collected, and
appends its `parts` onto our list of `parts`:

    else if (close) {
        let nested = molecule(stack.pop().formula).multiply(count);
        _parts = _parts.concat(nested.parts());
    }

Now we're properly processing sub-expressions and our tests return to
green.