# Alternate Groupings and Bug Fixes
Last up on our feature list is the ability to use parentheses and curly
brackets in addition to square brackets as sub-expression dividers.

We can add a test for this:

    expect(parseMolecule("K4{ON(SO3)2}2")).to.deep.equal({K: 4, O: 14, N: 2, S: 4});

The most straight-forward way to accomplish this is to just replace all
`"{"` and `"("` characters with `"["`, and `"}"` and `")"` characters
with `"]"` in our `formula`:

    formula.replace(/[{(]/g, "[").replace(/[})]/g, "]");

Unfortunately, this test also exposed another bug in our solution. it
looks like counts on sub-expressions aren't being applied to other
nested expressions.

We can fix this by giving our `stack` a bit more state. In addition to
building up our current expressions's `formula` as we go, we also want
to keep track of any nested `molecules` we encounter, so we can multiply
them by our `count`.

After making this change, all of our tests pass.